code/deployment/api/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.0
joblib>=1.0.0
numpy>=1.21.0
python-multipart>=0.0.5
